pyproject.toml
[project]
name = "openbase"
version = "0.1.0"
description = "Openbase Django Meta-Server"
authors = [{ name = "Gabe Montague", email = "gabemontague@outlook.com" }]
dependencies = [
    "Django==4.2.0",
    "djangorestframework==3.14.0",
    "django-cors-headers==4.0.0",
    "python-dotenv==1.1.1",
    "click==8.2.1",
    "gunicorn==21.0.0",
    "uvicorn==0.23.0",
    "requests==2.32.4",
    "djangorestframework-dataclasses==1.4.0",
    "whitenoise==6.9.0",
    "channels==4.2.2",
    "channels-redis==4.3.0",
]

[project.scripts]
openbase = "openbase.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["openbase"]
include = ["openbase/static/**/*"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
